% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedly-categories.R
\name{feedly_categories}
\alias{feedly_categories}
\title{Show Feedly Categories}
\usage{
feedly_categories(feedly_token = feedly_access_token())
}
\arguments{
\item{feedly_token}{Your Feedly Developer Access Token (see [feedly_access_token()])}
}
\description{
Utility function to just return category names and ids from your
Feedly collections. Authentication is _required_.
}
