% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedly-collections.R
\name{feedly_collections}
\alias{feedly_collections}
\title{Retrieve Feedly Connections}
\usage{
feedly_collections(with_stats = FALSE,
  feedly_token = feedly_access_token())
}
\arguments{
\item{with_stats}{if \code{TRUE}, return reading and tag stats for the past 31 days (default: \code{FALSE})}

\item{feedly_token}{Your Feedly Developer Access Token (see \code{\link[=feedly_access_token]{feedly_access_token()}})}
}
\value{
a data frame of all non-empty collections
}
\description{
Retrieve Feedly Connections
}
\references{
(\url{https://developer.feedly.com/v3/collections//})
}
