% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedly-enterprise-profile.R
\name{feedly_enterprise_profile}
\alias{feedly_enterprise_profile}
\title{Retrieve Your Feedly Enterprise Profile}
\usage{
feedly_enterprise_profile(feedly_token = feedly_access_token())
}
\arguments{
\item{feedly_token}{Your Feedly Developer Access Token (see \code{\link[=feedly_access_token]{feedly_access_token()}})}
}
\description{
Retrieve Your Feedly Enterprise Profile
}
\references{
(\url{https://developer.feedly.com/v3/opml/})
}
