% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedly-entry.R
\name{feedly_entry}
\alias{feedly_entry}
\title{Get the content of an entry}
\usage{
feedly_entry(entry_id, feedly_token = feedly_access_token())
}
\arguments{
\item{entry_id}{the unique, immutable ID for this particular article.}

\item{feedly_token}{Your Feedly Developer Access Token (see \code{\link[=feedly_access_token]{feedly_access_token()}})}
}
\description{
An entry is the atomic unit of content in the feedly cloud.
}
\references{
(\url{https://developer.feedly.com/v3/entries/})
}
