% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedly-feed-meta.R
\name{feedly_feed_meta}
\alias{feedly_feed_meta}
\title{Retrieve Metadata for a Feed}
\usage{
feedly_feed_meta(feed)
}
\arguments{
\item{feed}{an RSS feed URL. \code{feed/} will be prepended if not present.}
}
\value{
data frame (tibble) with the following fields:
\itemize{
\item \code{id}:
string the unique, immutable id of this feed.
\item \code{feedId}:
string same as id; for backward compatibility
\item \code{subscribers}:
integer number of feedly cloud subscribers who have this feed in their subscription list.
\item \code{title}:
string the feed name.
\item \code{description}:
Optional string the feed description.
\item \code{language}:
Optional string this field is a combination of the language reported by the RSS feed, and the language automatically detected from the feed’s content. It might not be accurate, as many feeds misreport it.
\item \code{velocity}:
Optional float the average number of articles published weekly. This number is updated every few days.
\item \code{website}:
Optional **url ** the website for this feed.
\item \code{topics}:
Optional string array an array of topics this feed covers. This list can be used in searches and mixes to build a list of related feeds and articles. E.g. if the list contains “productivity”, querying “productivity” in feed search will produce a list of related feeds.
\item \code{status}:
Optional string only returned if the feed cannot be polled. Values include “dead” (cannot be polled), “dead.flooded” (if the feed produces too many articles per day) and “dead.dropped” if the feed has been removed.
}
}
\description{
Retrieve Metadata for a Feed
}
\note{
This endpoint will do a real-time fetch of the feed if the submitted feed is not yet part of the feedly cloud.
}
\examples{
feedly_feed_meta("https://feeds.feedburner.com/rweeklylive")
}
\references{
(\url{https://developer.feedly.com/v3/feeds/})
}
