% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedly-opml.R
\name{feedly_opml}
\alias{feedly_opml}
\title{Retrieve Your Feedly OPML File}
\usage{
feedly_opml(as = c("text", "parsed"), feedly_token = feedly_access_token())
}
\arguments{
\item{as}{one of "\code{text}" (plain character XML) or "\code{parsed}" (which will return
an \code{xml2::xml_document})}

\item{feedly_token}{Your Feedly Developer Access Token (see \code{\link[=feedly_access_token]{feedly_access_token()}})}
}
\value{
character or \code{xml_document} depending on \code{as}
}
\description{
OPML (Outline Processor Markup Language) files are XML documents holding
"outlines". Specifically (for the context of Feedly) OPML files fully
describe the set of collections of RSS feeds, providing all the necessary
metadata to transfer between feed readers.
}
\references{
(\url{https://developer.feedly.com/v3/opml/})
}
