% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedly-search.R
\name{feedly_search}
\alias{feedly_search}
\title{Find feeds based on title, url or \code{#topic}}
\usage{
feedly_search(query, count = 20L, locale = NULL,
  feedly_token = feedly_access_token())
}
\arguments{
\item{query}{a full or partial title string, URL, or \code{#topic}}

\item{count}{number of items to return (API default is 20)}

\item{locale}{if not \code{NULL} then a Feedly-recognized locale string (see
\code{References}) to provide a hint to the search engine to return feeds
in that locale.}

\item{feedly_token}{Your Feedly Developer Access Token (see \code{\link[=feedly_access_token]{feedly_access_token()}})}
}
\value{
list similar to \code{\link[=feedly_stream]{feedly_stream()}}
}
\description{
Authorization is \emph{optional} for this API call. Pass in \code{NULL} to \code{feedly_token}
to make unauthenticated calls to this API endpoint.
}
\examples{
feedly_search("data science")
}
\references{
(\url{https://developer.feedly.com/v3/search/})
}
