% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedly-search-contents.R
\name{feedly_search_contents}
\alias{feedly_search_contents}
\title{Search content of a stream}
\usage{
feedly_search_contents(
  query,
  stream_id = NULL,
  fields = "all",
  embedded = NULL,
  engagement = NULL,
  count = 20L,
  locale = NULL,
  feedly_token = feedly_access_token()
)
}
\arguments{
\item{query}{a full or partial title string, URL, or \verb{#topic}}

\item{stream_id}{the id of the stream; a feed id, category id, tag id or a
system collection/category ids can be used as
stream ids. If \code{NULL} (the default) the server will use the
“\code{global.all}” (see \link{global_resource_ids}) collection/category.}

\item{fields}{if not "\code{all}" then a character vector of fields to use for
matching. SUpported fields are "\code{title}", "\code{author}", and "\code{keywords}".}

\item{embedded}{if not \code{NULL} then one of "\code{audio}", "\code{video}", "\code{doc}" or "\code{any}".
Using this parameter will limit results to also include this media type.
“\code{any}” means the article \emph{must} contain at least one embed.
Default behavior (i.e. \code{NULL}) is to not filter by embedded.}

\item{engagement}{if not \code{NULL} then either "\code{medium}" or "\code{high}".
Using this parameter will limit results to articles that have the
specified engagement. Default behavior (i.e. \code{NULL}) is to not
filter by engagement.}

\item{count}{number of items to return (max 20 for "pro" users)}

\item{locale}{if not \code{NULL} then a Feedly-recognized locale string (see
\code{References}) to provide a hint to the search engine to return feeds
in that locale.}

\item{feedly_token}{Your Feedly Developer Access Token (see \code{\link[=feedly_access_token]{feedly_access_token()}})}
}
\value{
list with a data frame element of \code{results}
}
\description{
Search content of a stream
}
\examples{
feedly_search_contents("data science")
}
\references{
(\url{https://developer.feedly.com/v3/search/}) & \href{https://feedly.uservoice.com/knowledgebase/articles/441699-power-search-tutorial}{Search Tutorial}
}
\seealso{
feedly_search_title
}
