% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedly-subscribe.R
\name{feedly_subscribe}
\alias{feedly_subscribe}
\title{Subscribe to an RSS feed}
\usage{
feedly_subscribe(feed_url, categories = NULL,
  feedly_token = feedly_access_token())
}
\arguments{
\item{feed_url}{a URL to an RSS feed (this will be validated).}

\item{categories}{if \code{NULL} then the feed will be assigned to the global (uncategorized) category.
Otherwise pass in a character vector of valid category names (use the
utility function \code{\link[=feedly_categories]{feedly_categories()}} to get a list of category names)}

\item{feedly_token}{Your Feedly Developer Access Token (see \code{\link[=feedly_access_token]{feedly_access_token()}})}
}
\description{
Authorization is \emph{required} for this API call.
}
\examples{
feedly_subscribe("https://journal.r-project.org/rss.atom")
}
\references{
(\url{https://developer.feedly.com/v3/subscriptions/#subscribe-to-a-feed/})
}
