% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedly-subscriptions.R
\name{feedly_subscriptions}
\alias{feedly_subscriptions}
\title{Retrieve Feedly Subscriptions}
\usage{
feedly_subscriptions(feedly_token = feedly_access_token())
}
\arguments{
\item{feedly_token}{Your Feedly Developer Access Token (see \code{\link[=feedly_access_token]{feedly_access_token()}})}
}
\value{
a data frame of feed subscriptions
}
\description{
This call \emph{requires} authentication.
}
\references{
(\url{https://developer.feedly.com/v3/subscriptions/})
}
