% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedly-tags.R
\name{feedly_tags}
\alias{feedly_tags}
\title{Retrieve List of Tags}
\usage{
feedly_tags(feedly_token = feedly_access_token())
}
\arguments{
\item{feedly_token}{Your Feedly Developer Access Token (see \code{\link[=feedly_access_token]{feedly_access_token()}})}
}
\value{
character or \code{xml_document} depending on \code{as}
}
\description{
This endpoint is useful for retrieving ids of Feedly Boards which would
then enable you to, say, make your own newsletters without upgrading to Teams.
}
\references{
(\url{https://developer.feedly.com/v3/tags/})
}
