% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-stream.R
\name{render_stream}
\alias{render_stream}
\title{Render a Feedly Stream Data Frame to RMarkdown}
\usage{
render_stream(
  feedly_stream,
  title = "Feedly Stream Summary",
  include_visual = FALSE,
  browse = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{feedly_stream}{output from \code{\link[=feedly_stream]{feedly_stream()}}. Ideally, you'll filter
this so it has a minimal number of elements (unless you like scrolling
through large HTML documents)}

\item{title}{A title for your "report"}

\item{include_visual}{if \code{TRUE} then any \code{visual_url} image content will
be rendered with the summary text. Default is \code{FALSE}.}

\item{browse}{if \code{TRUE} open the HTML document for viewing. Default is \code{FALSE}}

\item{quiet}{passed on to \code{\link[rmarkdown:render]{rmarkdown::render()}}. Default is \code{TRUE}.}
}
\value{
a list with the contents of the R Markdown document and the
rendered HTML file (so you can save/tweak before publishing).
}
\description{
This function takes a minimalist approach to formatting an R Markdown
document from the contents of a Feedly stream. It uses the feed item
title, URL (first \code{alternate} one), published data and summary content to make
each entry.
}
\note{
This function requires that \code{htmltools}, \code{rmarkdown} and \code{glue} are installed
on your system.
}
