% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seymour-package.R
\docType{package}
\name{seymour}
\alias{seymour}
\alias{seymour-package}
\title{Tools to Work with the 'Feedly' 'API'}
\description{
'Feedly' is a news aggregator application for various web browsers and
mobile devices running 'iOS' and 'Android', also available as a cloud-based
service. It compiles news feeds from a variety of online sources for the
user to customize and share with others. Methods are provided to retrieve
information about and contents of 'Feedly' collections and streams.
}
\details{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/seymour}
\item BugReports: \url{https://gitlab.com/hrbrmstr/seymour/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
