% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedly-profile.R
\name{feedly_profile}
\alias{feedly_profile}
\title{Retrieve Your Feedly Profile}
\usage{
feedly_profile(feedly_token = feedly_access_token())
}
\arguments{
\item{feedly_token}{Your Feedly Developer Access Token (see \code{\link[=feedly_access_token]{feedly_access_token()}})}
}
\description{
Retrieve Your Feedly Profile
}
\references{
(\url{https://developer.feedly.com/v3/opml/})
}
