% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seymour-package.R
\docType{package}
\name{seymour}
\alias{seymour}
\alias{seymour-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/seymour}
\item BugReports: \url{https://gitlab.com/hrbrmstr/seymour/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
